+++ cmd/hidden_lake/messenger/internal/database/database.go
@@ -3,6 +3,7 @@ package database
 import (
 	"sync"
 
+	"github.com/number571/go-peer/cmd/hidden_lake/messenger/pkg/settings"
 	"github.com/number571/go-peer/pkg/encoding"
 	"github.com/number571/go-peer/pkg/errors"
 	"github.com/number571/go-peer/pkg/storage"
@@ -14,12 +15,13 @@ type sKeyValueDB struct {
 	fDB    *gp_database.IKVDatabase
 }
 
-func NewKeyValueDB(pPath string, pKey []byte) (IKVDatabase, error) {
+func NewKeyValueDB(pPath, pPassword string) (IKVDatabase, error) {
 	db, err := gp_database.NewKeyValueDB(
 		storage.NewSettings(&storage.SSettings{
-			FPath:      pPath,
-			FHashing:   true,
-			FCipherKey: pKey,
+			FPath:     pPath,
+			FHashing:  true,
+			FWorkSize: settings.CWorkForKeys,
+			FPassword: pPassword,
 		}),
 	)
 	if err != nil {