+++ cmd/hidden_lake/service/pkg/app/init_database.go
@@ -12,9 +12,10 @@ import (
 func (p *sApp) initDatabase() error {
 	db, err := database.NewKeyValueDB(
 		storage.NewSettings(&storage.SSettings{
-			FPath:      fmt.Sprintf("%s/%s", p.fPathTo, pkg_settings.CPathDB),
-			FHashing:   false,
-			FCipherKey: []byte("_"),
+			FPath:     fmt.Sprintf("%s/%s", p.fPathTo, pkg_settings.CPathDB),
+			FHashing:  false,
+			FWorkSize: 1,
+			FPassword: "_",
 		}),
 	)
 	if err != nil {