+++ cmd/hidden_lake/traffic/internal/database/database.go
@@ -26,9 +26,10 @@ func NewKeyValueDB(pSett ISettings) (IKVDatabase, error) {
 
 	kvDB, err := database.NewKeyValueDB(
 		storage.NewSettings(&storage.SSettings{
-			FPath:      pSett.GetPath(),
-			FHashing:   false,
-			FCipherKey: []byte("_"),
+			FPath:     pSett.GetPath(),
+			FHashing:  false,
+			FWorkSize: 1,
+			FPassword: "_",
 		}),
 	)
 	if err != nil {