+++ pkg/crypto/entropy/types.go
@@ -1,5 +1,0 @@
-package entropy
-
-type IEntropyBooster interface {
-	BoostEntropy([]byte) []byte
-}