+++ pkg/crypto/keybuilder/key_builder.go
@@ -1,0 +1,34 @@
+package keybuilder
+
+import (
+	"crypto/sha256"
+
+	"github.com/number571/go-peer/pkg/crypto/symmetric"
+	"golang.org/x/crypto/pbkdf2"
+)
+
+var (
+	_ IKeyBuilder = &sKeyBuilder{}
+)
+
+type sKeyBuilder struct {
+	fSalt []byte
+	fBits uint64
+}
+
+func NewKeyBuilder(pBits uint64, pSalt []byte) IKeyBuilder {
+	return &sKeyBuilder{
+		fBits: pBits,
+		fSalt: pSalt,
+	}
+}
+
+func (p *sKeyBuilder) Build(pPassword []byte) []byte {
+	return pbkdf2.Key(
+		pPassword,
+		p.fSalt,
+		(1 << p.fBits),
+		int(symmetric.CAESKeySize),
+		sha256.New,
+	)
+}