+++ pkg/crypto/keybuilder/types.go
@@ -1,0 +1,5 @@
+package keybuilder
+
+type IKeyBuilder interface {
+	Build([]byte) []byte
+}