+++ pkg/storage/settings.go
@@ -6,16 +6,18 @@ var (
 
 type SSettings sSettings
 type sSettings struct {
-	FPath      string
-	FHashing   bool
-	FCipherKey []byte
+	FPath     string
+	FHashing  bool
+	FWorkSize uint64
+	FPassword string
 }
 
 func NewSettings(pSett *SSettings) ISettings {
 	return (&sSettings{
-		FPath:      pSett.FPath,
-		FHashing:   pSett.FHashing,
-		FCipherKey: pSett.FCipherKey,
+		FPath:     pSett.FPath,
+		FHashing:  pSett.FHashing,
+		FWorkSize: pSett.FWorkSize,
+		FPassword: pSett.FPassword,
 	}).mustNotNull()
 }
 
@@ -23,8 +25,8 @@ func (p *sSettings) mustNotNull() ISettings {
 	if p.FPath == "" {
 		panic(`p.FPath == ""`)
 	}
-	if p.FCipherKey == nil {
-		panic(`p.FCipherKey == nil`)
+	if p.FWorkSize == 0 {
+		panic(`p.FWorkSize == 0`)
 	}
 	return p
 }
@@ -37,6 +39,10 @@ func (p *sSettings) GetHashing() bool {
 	return p.FHashing
 }
 
-func (p *sSettings) GetCipherKey() []byte {
-	return p.FCipherKey
+func (p *sSettings) GetWorkSize() uint64 {
+	return p.FWorkSize
+}
+
+func (p *sSettings) GetPassword() string {
+	return p.FPassword
 }